chore: move plan calculator inside change plan
