+++ src/App.js
@@ -22,7 +22,7 @@ import Offline from './components/Offline/Offline';
 import PushNotifications from './components/PushNotifications/PushNotifications';
 import SubscribersLegacyUrlRedirect from './components/Reports/Subscribers/SubscribersLegacyUrlRedirect';
 import ChangePlan from './components/ChangePlan/ChangePlan';
-import PlanCalculator from './components/DopplerPlus/PlanCalculator';
+import PlanCalculator from './components/ChangePlan/PlanCalculator';
 import AgenciesForm from './components/DopplerPlus/AgenciesForm';
 
 /**
